made requested pr changes
