+++ client/src/styles/CustomEventStyles.tsx
@@ -13,3 +13,10 @@ export const StyledListItemText = styled(ListItemText)`
 export const ColourButton = styled(Button)`
   text-transform: none;
 `;
+
+export const ExecuteButton = styled(Button)`
+  margin-top: 4px;
+  height: 40px;
+  width: 100%;
+  border-radius: 0px 0px 5px 5px;
+`;
